src/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dateparser==1.1.8
filelock==3.12.0
Flask==2.3.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nodeenv==1.8.0
numpy==1.24.3
pandas==2.0.1
prisma==0.8.2
pycryptodome==3.18.0
pydantic==1.10.8
python-binance==1.0.17
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sympy==1.12
threadpoolctl==3.1.0
tomlkit==0.11.8
torch==2.0.1
typing_extensions==4.6.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.7.0
urllib3==2.0.2
websockets==11.0.3
Werkzeug==2.3.4
yarl==1.9.2
